rename SqliteMemoryStore to SQLiteMemoryStore
