+++ java/connectors/semantickernel-connectors-memory-sqlite/src/main/java/com/microsoft/semantickernel/connectors/memory/sqlite/Database.java
@@ -4,16 +4,15 @@ package com.microsoft.semantickernel.connectors.memory.sqlite;
 import com.microsoft.semantickernel.memory.DataEntryBase;
 import com.microsoft.semantickernel.memory.MemoryException;
 import com.microsoft.semantickernel.memory.MemoryException.ErrorCodes;
-import org.sqlite.SQLiteErrorCode;
-import org.sqlite.SQLiteException;
-import reactor.core.publisher.Mono;
-import reactor.core.scheduler.Schedulers;
-
 import java.sql.*;
 import java.time.ZonedDateTime;
 import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
 import java.util.List;
+import org.sqlite.SQLiteErrorCode;
+import org.sqlite.SQLiteException;
+import reactor.core.publisher.Mono;
+import reactor.core.scheduler.Schedulers;
 
 public class Database {
 
@@ -21,7 +20,8 @@ public class Database {
         private final String metadata;
         private final String embedding;
 
-        public DatabaseEntry(String key, String metadata, String embedding, ZonedDateTime timestamp) {
+        public DatabaseEntry(
+                String key, String metadata, String embedding, ZonedDateTime timestamp) {
             super(key, timestamp);
             this.metadata = metadata;
             this.embedding = embedding;
@@ -34,77 +34,81 @@ public class Database {
         public String getEmbedding() {
             return embedding;
         }
-
     }
 
     // Convenience method to format a ZonedDateTime in a format acceptable to SQLite
     private static String formatDatetime(ZonedDateTime datetime) {
-      if (datetime == null) return "";
-      return datetime.format(DateTimeFormatter.ISO_OFFSET_DATE_TIME);
+        if (datetime == null) return "";
+        return datetime.format(DateTimeFormatter.ISO_OFFSET_DATE_TIME);
     }
 
     // Convenience method to parse a SQLite datetime string into a ZonedDateTime
     private static ZonedDateTime parseDatetime(String datetime) {
-      if (datetime == null || datetime.isEmpty()) return null;
+        if (datetime == null || datetime.isEmpty()) return null;
         return ZonedDateTime.parse(datetime, DateTimeFormatter.ISO_OFFSET_DATE_TIME);
     }
 
-  private static final String COLLECTIONS_TABLE_NAME = "SKCollectionTable";
-  private static final String TABLE_NAME = "SKMemoryTable";
-  private static final String INDEX_NAME = "SKMemoryIndex";
+    private static final String COLLECTIONS_TABLE_NAME = "SKCollectionTable";
+    private static final String TABLE_NAME = "SKMemoryTable";
+    private static final String INDEX_NAME = "SKMemoryIndex";
 
     public Mono<Void> createTableAsync(Connection connection) {
         return Mono.fromRunnable(
                         () -> {
-                          String createCollectionKeyTable =
-                              "CREATE TABLE IF NOT EXISTS "
-                                  + COLLECTIONS_TABLE_NAME
-                                  + " ("
-                                  + "id INTEGER PRIMARY KEY, "
-                                  + "name TEXT NOT NULL UNIQUE"
-                                  + " )";
+                            String createCollectionKeyTable =
+                                    "CREATE TABLE IF NOT EXISTS "
+                                            + COLLECTIONS_TABLE_NAME
+                                            + " ("
+                                            + "id INTEGER PRIMARY KEY, "
+                                            + "name TEXT NOT NULL UNIQUE"
+                                            + " )";
 
-                          String createSKMemoryTable =
-                              "CREATE TABLE IF NOT EXISTS "
-                                  + TABLE_NAME
-                                  + " ("
-                                  + "collection INTEGER, "
-                                  + "key TEXT, "
-                                  + "metadata TEXT, "
-                                  + "embedding TEXT, "
-                                  + "timestamp TEXT, "
-                                  + "FOREIGN KEY(collection) REFERENCES "
-                                  + COLLECTIONS_TABLE_NAME + "(id)"
-                                  + " )";
+                            String createSKMemoryTable =
+                                    "CREATE TABLE IF NOT EXISTS "
+                                            + TABLE_NAME
+                                            + " ("
+                                            + "collection INTEGER, "
+                                            + "key TEXT, "
+                                            + "metadata TEXT, "
+                                            + "embedding TEXT, "
+                                            + "timestamp TEXT, "
+                                            + "FOREIGN KEY(collection) REFERENCES "
+                                            + COLLECTIONS_TABLE_NAME
+                                            + "(id)"
+                                            + " )";
 
-                          String createIndex =
-                              "CREATE INDEX " + INDEX_NAME
-                                  + " ON " + TABLE_NAME + "(collection)";
+                            String createIndex =
+                                    "CREATE INDEX "
+                                            + INDEX_NAME
+                                            + " ON "
+                                            + TABLE_NAME
+                                            + "(collection)";
 
-                          try (Statement statement = connection.createStatement()) {
-                            statement.addBatch(createCollectionKeyTable);
-                            statement.addBatch(createSKMemoryTable);
-                            statement.addBatch(createIndex);
-                            statement.executeBatch();
-                          } catch (SQLException e) {
-                            throw new SQLConnectorException("\"CREATE TABLE\" failed", e);
-                          }
+                            try (Statement statement = connection.createStatement()) {
+                                statement.addBatch(createCollectionKeyTable);
+                                statement.addBatch(createSKMemoryTable);
+                                statement.addBatch(createIndex);
+                                statement.executeBatch();
+                            } catch (SQLException e) {
+                                throw new SQLConnectorException("\"CREATE TABLE\" failed", e);
+                            }
                         })
                 .subscribeOn(Schedulers.boundedElastic())
                 .then();
     }
 
     public Mono<Void> createCollectionAsync(Connection connection, String collectionName) {
-          return Mono.fromRunnable(
+        return Mono.fromRunnable(
                         () -> {
-                            String query = "INSERT INTO " + COLLECTIONS_TABLE_NAME + " (name) VALUES (?)";
+                            String query =
+                                    "INSERT INTO " + COLLECTIONS_TABLE_NAME + " (name) VALUES (?)";
                             try (PreparedStatement statement = connection.prepareStatement(query)) {
                                 statement.setString(1, collectionName);
                                 statement.executeUpdate();
                             } catch (SQLiteException e) {
-                              if (e.getResultCode() != SQLiteErrorCode.SQLITE_CONSTRAINT_UNIQUE) {
-                                throw new SQLConnectorException("\"INSERT INTO\" failed", e);
-                              }
+                                if (e.getResultCode() != SQLiteErrorCode.SQLITE_CONSTRAINT_UNIQUE) {
+                                    throw new SQLConnectorException("\"INSERT INTO\" failed", e);
+                                }
                             } catch (SQLException e) {
                                 throw new SQLConnectorException("\"INSERT INTO\" failed", e);
                             }
@@ -127,7 +131,9 @@ public class Database {
                                     "UPDATE "
                                             + TABLE_NAME
                                             + " SET metadata = ?, embedding = ?, timestamp = ?"
-                                            + " WHERE collection = (SELECT id FROM " + COLLECTIONS_TABLE_NAME + " WHERE name = ?)"
+                                            + " WHERE collection = (SELECT id FROM "
+                                            + COLLECTIONS_TABLE_NAME
+                                            + " WHERE name = ?)"
                                             + " AND key = ?";
                             try (PreparedStatement statement = connection.prepareStatement(query)) {
                                 statement.setString(1, metadata != null ? metadata : "");
@@ -157,7 +163,9 @@ public class Database {
                                     "INSERT OR IGNORE INTO "
                                             + TABLE_NAME
                                             + " (collection, key, metadata, embedding, timestamp)"
-                                            + " VALUES ((SELECT id FROM " + COLLECTIONS_TABLE_NAME + " WHERE name = ?), ?, ?, ?, ?)";
+                                            + " VALUES ((SELECT id FROM "
+                                            + COLLECTIONS_TABLE_NAME
+                                            + " WHERE name = ?), ?, ?, ?, ?)";
                             try (PreparedStatement statement = connection.prepareStatement(query)) {
                                 statement.setString(1, collection);
                                 statement.setString(2, key);
@@ -166,7 +174,8 @@ public class Database {
                                 statement.setString(5, formatDatetime(timestamp));
                                 statement.executeUpdate();
                             } catch (SQLException e) {
-                                throw new SQLConnectorException("\"INSERT OR IGNORE INTO\" failed", e);
+                                throw new SQLConnectorException(
+                                        "\"INSERT OR IGNORE INTO\" failed", e);
                             }
                         })
                 .subscribeOn(Schedulers.boundedElastic())
@@ -182,7 +191,7 @@ public class Database {
             throws SQLException {
         String query = "SELECT name FROM " + COLLECTIONS_TABLE_NAME + " WHERE name = ?";
         try (PreparedStatement statement = connection.prepareStatement(query)) {
-          statement.setString(1, collectionName);
+            statement.setString(1, collectionName);
             ResultSet resultSet = statement.executeQuery();
             return resultSet.next();
         }
@@ -200,7 +209,8 @@ public class Database {
                                     collections.add(collection);
                                 }
                             } catch (SQLException e) {
-                                return Mono.error(new SQLConnectorException("\"SELECT\" failed", e));
+                                return Mono.error(
+                                        new SQLConnectorException("\"SELECT\" failed", e));
                             }
                             return Mono.just(collections);
                         })
@@ -211,9 +221,13 @@ public class Database {
         return Mono.defer(
                         () -> {
                             List<DatabaseEntry> entries = new ArrayList<>();
-                            String query = "SELECT * FROM " + TABLE_NAME
-                                + " WHERE collection ="
-                                + " (SELECT id FROM " + COLLECTIONS_TABLE_NAME + " WHERE name = ?)";
+                            String query =
+                                    "SELECT * FROM "
+                                            + TABLE_NAME
+                                            + " WHERE collection ="
+                                            + " (SELECT id FROM "
+                                            + COLLECTIONS_TABLE_NAME
+                                            + " WHERE name = ?)";
                             try (PreparedStatement statement = connection.prepareStatement(query)) {
                                 statement.setString(1, collectionName);
                                 ResultSet resultSet = statement.executeQuery();
@@ -224,10 +238,12 @@ public class Database {
                                     String timestamp = resultSet.getString("timestamp");
                                     ZonedDateTime zonedDateTime = parseDatetime(timestamp);
                                     entries.add(
-                                            new DatabaseEntry(key, metadata, embedding, zonedDateTime));
+                                            new DatabaseEntry(
+                                                    key, metadata, embedding, zonedDateTime));
                                 }
                             } catch (SQLException e) {
-                                return Mono.error(new SQLConnectorException("\"SELECT * FROM\" failed", e));
+                                return Mono.error(
+                                        new SQLConnectorException("\"SELECT * FROM\" failed", e));
                             }
                             return Mono.just(entries);
                         })
@@ -237,10 +253,14 @@ public class Database {
     public Mono<DatabaseEntry> readAsync(Connection connection, String collectionName, String key) {
         return Mono.defer(
                         () -> {
-                          String query = "SELECT * FROM " + TABLE_NAME
-                              + " WHERE collection ="
-                              + " (SELECT id FROM " + COLLECTIONS_TABLE_NAME + " WHERE name = ?)"
-                              + " AND key = ?";
+                            String query =
+                                    "SELECT * FROM "
+                                            + TABLE_NAME
+                                            + " WHERE collection ="
+                                            + " (SELECT id FROM "
+                                            + COLLECTIONS_TABLE_NAME
+                                            + " WHERE name = ?)"
+                                            + " AND key = ?";
                             try (PreparedStatement statement = connection.prepareStatement(query)) {
                                 statement.setString(1, collectionName);
                                 statement.setString(2, key != null && !key.isEmpty() ? key : null);
@@ -251,10 +271,12 @@ public class Database {
                                     String timestamp = resultSet.getString("timestamp");
                                     ZonedDateTime zonedDateTime = parseDatetime(timestamp);
                                     return Mono.just(
-                                            new DatabaseEntry(key, metadata, embedding, zonedDateTime));
+                                            new DatabaseEntry(
+                                                    key, metadata, embedding, zonedDateTime));
                                 }
                             } catch (SQLException e) {
-                                return Mono.error(new SQLConnectorException("\"SELECT * FROM\" failed", e));
+                                return Mono.error(
+                                        new SQLConnectorException("\"SELECT * FROM\" failed", e));
                             }
                             return Mono.empty();
                         })
@@ -264,17 +286,24 @@ public class Database {
     public Mono<Void> deleteCollectionAsync(Connection connection, String collectionName) {
         return Mono.fromRunnable(
                         () -> {
-                            String query1 = "DELETE FROM " + TABLE_NAME
-                                + " WHERE collection ="
-                                + " (SELECT id FROM " + COLLECTIONS_TABLE_NAME + " WHERE name = ?)";
-                            String query2 = "DELETE FROM " + COLLECTIONS_TABLE_NAME + " WHERE name = ?";
+                            String query1 =
+                                    "DELETE FROM "
+                                            + TABLE_NAME
+                                            + " WHERE collection ="
+                                            + " (SELECT id FROM "
+                                            + COLLECTIONS_TABLE_NAME
+                                            + " WHERE name = ?)";
+                            String query2 =
+                                    "DELETE FROM " + COLLECTIONS_TABLE_NAME + " WHERE name = ?";
                             try (PreparedStatement statement = connection.prepareStatement(query1);
-                                 PreparedStatement statement2 = connection.prepareStatement(query2)) {
+                                    PreparedStatement statement2 =
+                                            connection.prepareStatement(query2)) {
                                 statement.setString(1, collectionName);
                                 statement.executeUpdate();
                                 statement2.setString(1, collectionName);
                                 if (statement2.executeUpdate() == 0) {
-                                    throw new MemoryException(ErrorCodes.ATTEMPTED_TO_ACCESS_NONEXISTENT_COLLECTION);
+                                    throw new MemoryException(
+                                            ErrorCodes.ATTEMPTED_TO_ACCESS_NONEXISTENT_COLLECTION);
                                 }
                             } catch (SQLException e) {
                                 throw new SQLConnectorException("\"DELETE FROM\" failed", e);
@@ -287,10 +316,14 @@ public class Database {
     public Mono<Void> deleteAsync(Connection connection, String collectionName, String key) {
         return Mono.fromRunnable(
                         () -> {
-                            String query = "DELETE FROM " + TABLE_NAME
-                                + " WHERE collection ="
-                                + " (SELECT id FROM " + COLLECTIONS_TABLE_NAME + " WHERE name = ?)"
-                                + " AND key = ?";
+                            String query =
+                                    "DELETE FROM "
+                                            + TABLE_NAME
+                                            + " WHERE collection ="
+                                            + " (SELECT id FROM "
+                                            + COLLECTIONS_TABLE_NAME
+                                            + " WHERE name = ?)"
+                                            + " AND key = ?";
                             try (PreparedStatement statement = connection.prepareStatement(query)) {
                                 statement.setString(1, collectionName);
                                 statement.setString(2, key != null && !key.isEmpty() ? key : null);
@@ -306,10 +339,14 @@ public class Database {
     public Mono<Void> deleteEmptyAsync(Connection connection, String collectionName) {
         return Mono.fromRunnable(
                         () -> {
-                          String query = "DELETE FROM " + TABLE_NAME
-                              + " WHERE collection ="
-                              + " (SELECT id FROM " + COLLECTIONS_TABLE_NAME + " WHERE name = ?)"
-                              + " AND key is NULL";
+                            String query =
+                                    "DELETE FROM "
+                                            + TABLE_NAME
+                                            + " WHERE collection ="
+                                            + " (SELECT id FROM "
+                                            + COLLECTIONS_TABLE_NAME
+                                            + " WHERE name = ?)"
+                                            + " AND key is NULL";
                             try (PreparedStatement statement = connection.prepareStatement(query)) {
                                 statement.setString(1, collectionName);
                                 statement.executeUpdate();
